tests: add decode
this may not stay in this repo but its useful for speed tests.
really though this just shows that we need a way of dumping
average performance metrics from the beamformer. There is too much
random overhead over in this path. (note that the beamformer
should also have CPU side profiling added)