core/build: go all in on asan
asan is now always enabled in debug builds (it can still be
optionally enabled in release build). Manually poisoning has some
drawbacks, particularly for arena_stream(), but it will catch
issues with freelist usage and anything in a frame up until a call
to arena_stream(). That part could be worked on later as necessary.