core: asynchronous RF upload thread
This significantly improves performance by overlapping data
uploads with compute. I'm not 100% sure about by AMD card but most
modern NVIDIA cards have an asynchronous upload queue that can be
run at the same time as the compute queue is being processed.
Apparently the only reliable way in OpenGL of getting the driver
to use that upload queue is to have a separate context that only
ever issues memory related commands, never drawing commands.

With some careful use of fences and a persistently mapped
oversized ssbo we can keep the compute thread processing a frame
while a new RF is being uploaded.

This increases the rf upload throughput on my weak desktop GPU
past what we used to get on the RTX 4090 before we cleaned up the
matlab side. This should effectively remove all synchronization
but it still needs to be profiled.