core/lib: support multiple parameter blocks
currently ui is limited to display block 0 and won't display
anything unless data was sent to that block. Additionally garbage
collection on blocks that appear to be unused is not currently
implemented. Therefore GPU data such as textures and UBOS
associated with the block are held until the program exits even
when the block has stopped being used.