build: add meta program
Currently this handles:
  - generating a list of shader variations with baked parameters
    - these are measured to be more performant than behaviour modification
      through values specified at runtime
  - a set of enums that are required by the shaders
    - some of which are global and used below
  - MATLAB bindings based on global enums
  - modified helper library header based on global enums

No optimization was performed on the "compiler" so it may get slow
as the meta file expands. A fair amount of work was done to ensure
that information that is shared between shader variations is not
duplicated over and over.

The generated C code is committed so that it is clear what parts
of the actual compiled program have changed when something in the
meta file or "compiler" is modified.