shaders/das: move iq rotation into if gaurd
shader compiler isn't smart enough to omit this when the if is
false and ends up evaluating sin and cos anyways. this gives a
minor performance boost when we are not beamforming beyond the end
of the data and a more substantial one if we are.

Linear is a special case which seems to run better if it always
has to evaluate the sin/cos