core/ui: hash beamformer bake parameters to avoid reloading unnecessarily
xxhash was chosen as its very fast, making using of vector
instructions on both x86_64 and aarch64. The file is huge because
its full of yapping but the actual code it compiles down to is
quite small. Its also single header so trivial to embed in the
existing single TU build.

Its also quite good for small inputs so it can also be used for
hashing short strings.