core: fix race condition in rf upload function
While the code prior to this commit contained a condition to
prevent the two helper threads from sleeping during live imaging
the imaging upload rate was still limited to 60 FPS. This was
because the upload thread was stuck in the loop until the main
thread allowed it to exit. Fixing this exposed a race in the
library upload code where the upload thread could preempt the
library into taking the scratch space lock before the library was
able to upload data.

To solve this I utilize the scratch_rf_size as a separate
synchronization point. If the value is 0 the upload thread will
not try to lock the scratch region. If its any other value the
upload thread sets it to zero and takes the lock (the library
cannot try to take the lock again until the upload thread releases
the UploadRF fence)

Furthermore I made a pass through all the places where values are
being modified across threads and ensured they are done
atomically; this is the only way to ensure that their value
becomes visible to other threads. All atomic operations for
GCC/Clang were switched to SEQ_CST; while some of them could
technically use weaker constraints, I do not want to think about
them until they become performance issues (which they probably
never will).