build/meta: support arbitrary binary data bakes, bake font data
generated asset source code is excluded for now as it is bigger
than source file. in the future it could just get directly linked
into the executable but that is more difficult to do in a platform
agnostic way

this should be the only other file that was required to be
available at runtime. now the binary should be self contained