shaders/filter: fix filter reversing
the filter is already time reversed before it is uploaded. it
shouldn't be time reversed again. this doesn't matter for the
normal low pass filter since its symmetric but the chirp filter is
not. This caused a shift in the image (proportional to the filter
length) and a loss of contrast.

There was also an off by one in the min_sample calculation for the
sample preloading. This has minimal effect on the image (1 sample
shift) but did mean that output sample n did not contain input
sample n.

Preloading loop was also reworked to be, in theory, friendlier to
the GPU in terms of LDS bank accesses but it has no effect on the
performance on my GPU