core/lib: always perform channel mapping on the CPU
There are no cases where this would be slower. To perform the
mem_copy the CPU must touch every single byte in the array. By
simply not copying channels that are 0 we get a latency reduction.
This also means that none of the shaders need to care about
channel mapping which simplifies everything.

In the common case where half of the channels are 0 this will
reduce the latency from memory copies by at least half. It will
also reduce the latency from uploading to the GPU by a similar
amount.

NOTE: nothing about the external API needs to change; the only way
the previous code could work was if the channel mapping was
uploaded prior to the data being sent. Therefore the library can
just load it and do the mapping.