os: os_take_lock: fix return behaviour for infinite timeouts
callers of this function expect that it will not return until the
lock is successfully acquired. this means that regardless of what
the underlying syscall returns it is never an error for an
infinite timeout.

I noticed this when the value would change between the atomic_cas
and entering the kernel.