wip: core: switch beamformer output from image/texture to ssbo
this is needed for an in progress optimization for DAS. it is also
the only way to support NPOT output in the Z dimension.

for now the min/max shader is broken, but we don't use it anyways.

TODO: test with lower power devices. if the glFinish() doesn't get
removed the performance with this tanks. But I think it was still
there to prevent lagging on other devices.