ui: add acquistion kind live control
If the live imaging parameters indicate that there are multiple
kinds of sequences programmed an additional menu will be enabled
allowing the user to switch between them. This is currently
implemented in a fairly hacky/inflexible way and can be greatly
improved with better imaging system integration in the future
(especially if the beamformer drives the programming).