add some code to allow building w32 application on linux
Since we define our own windows prototypes/structures we don't
actually need to windows headers to build the application. With
some help the application can be compiled on linux to make sure
any refactoring doesn't break the build. With these changes the
application can be built but will fail to link.

Note: '-fms-extensions' can be used on linux as well, and some of
the extensions are actually useful (Linux kernel is now using
them), so there is no harm on having it enabled by default.